Separated jetton and DAO test suites
