+++ tests/utils.ts
@@ -1,0 +1,44 @@
+
+import { Address, toNano, fromNano } from "ton";
+
+export const randomAddress = (wc: number = 0) => {
+    const buf = Buffer.alloc(32);
+    for (let i = 0; i < buf.length; i++) {
+        buf[i] = Math.floor(Math.random() * 256);
+    }
+    return new Address(wc, buf);
+};
+
+export const getRandom = (min:number, max:number) => {
+    return Math.random() * (max - min) + min;
+}
+
+enum roundMode {floor, ceil, round};
+
+export const getRandomInt = (min:number, max:number, mode: roundMode = roundMode.floor) => {
+    let res = getRandom(min, max);
+
+    if(mode == roundMode.floor) {
+        res = Math.floor(res);
+    }
+    else if(mode == roundMode.ceil) {
+        res = Math.ceil(res);
+    }
+    else {
+        res = Math.round(res);
+    }
+
+    return res;
+}
+
+export const getRandomTon = (min:number, max:number): bigint => {
+    return toNano(getRandom(min, max).toFixed(9));
+}
+
+export const getRandomExp = () => {
+    return BigInt(Math.floor(Date.now() / 1000) + getRandomInt(10, 100));
+}
+
+export const renewExp = (cur:bigint) => {
+    return Math.floor(Date.now() / 1000) >= cur ? getRandomExp() : cur;
+}
